fix(lsp): formatting should trim extra final newline
With multiple newlines at the end of a document the LSP formatter is
leaving an extra final newline instead of leaving just a single final
newline.

This is due to the compute_text_edits using incorrect position when
calculating where the end of the document is when it ends with a
newline.

The fix is to use the last line of the old text if the new text finishes
with a newline.

Issue: https://github.com/tombi-toml/tombi/issues/934